refactor(agents): extract provider model normalization
