+++ docs/refactor/cleanup.md
@@ -2,6 +2,6 @@
 
 - [x] Extract `models list` row/supplement helpers.
 - [x] Split `models list` forward-compat tests by concern.
-- [ ] Extract provider transport normalization from `pi-embedded-runner/model.ts`.
+- [x] Extract provider transport normalization from `pi-embedded-runner/model.ts`.
 - [ ] Split `ensureOpenClawModelsJson()` into planning + IO layers.
 - [ ] Split provider discovery helpers out of `models-config.providers.ts`.
