+++ src/agents/pi-embedded-runner/model.ts
@@ -6,10 +6,10 @@ import { resolveOpenClawAgentDir } from "../agent-paths.js";
 import { DEFAULT_CONTEXT_TOKENS } from "../defaults.js";
 import { buildModelAliasLines } from "../model-alias-lines.js";
 import { isSecretRefHeaderValueMarker } from "../model-auth-markers.js";
-import { normalizeModelCompat } from "../model-compat.js";
 import { resolveForwardCompatModel } from "../model-forward-compat.js";
 import { findNormalizedProviderValue, normalizeProviderId } from "../model-selection.js";
 import { discoverAuthStorage, discoverModels } from "../pi-model-discovery.js";
+import { normalizeResolvedProviderModel } from "./model.provider-normalization.js";
 
 type InlineModelEntry = ModelDefinitionConfig & {
   provider: string;
@@ -22,8 +22,6 @@ type InlineProviderConfig = {
   models?: ModelDefinitionConfig[];
   headers?: unknown;
 };
-
-const OPENAI_CODEX_BASE_URL = "https://chatgpt.com/backend-api";
 
 function sanitizeModelHeaders(
   headers: unknown,
@@ -45,58 +43,8 @@ function sanitizeModelHeaders(
   return Object.keys(next).length > 0 ? next : undefined;
 }
 
-function isOpenAIApiBaseUrl(baseUrl?: string): boolean {
-  const trimmed = baseUrl?.trim();
-  if (!trimmed) {
-    return false;
-  }
-  return /^https?:\/\/api\.openai\.com(?:\/v1)?\/?$/i.test(trimmed);
-}
-
-function isOpenAICodexBaseUrl(baseUrl?: string): boolean {
-  const trimmed = baseUrl?.trim();
-  if (!trimmed) {
-    return false;
-  }
-  return /^https?:\/\/chatgpt\.com\/backend-api\/?$/i.test(trimmed);
-}
-
-function normalizeOpenAICodexTransport(params: {
-  provider: string;
-  model: Model<Api>;
-}): Model<Api> {
-  if (normalizeProviderId(params.provider) !== "openai-codex") {
-    return params.model;
-  }
-
-  const useCodexTransport =
-    !params.model.baseUrl ||
-    isOpenAIApiBaseUrl(params.model.baseUrl) ||
-    isOpenAICodexBaseUrl(params.model.baseUrl);
-
-  const nextApi =
-    useCodexTransport && params.model.api === "openai-responses"
-      ? ("openai-codex-responses" as const)
-      : params.model.api;
-  const nextBaseUrl =
-    nextApi === "openai-codex-responses" &&
-    (!params.model.baseUrl || isOpenAIApiBaseUrl(params.model.baseUrl))
-      ? OPENAI_CODEX_BASE_URL
-      : params.model.baseUrl;
-
-  if (nextApi === params.model.api && nextBaseUrl === params.model.baseUrl) {
-    return params.model;
-  }
-
-  return {
-    ...params.model,
-    api: nextApi,
-    baseUrl: nextBaseUrl,
-  } as Model<Api>;
-}
-
 function normalizeResolvedModel(params: { provider: string; model: Model<Api> }): Model<Api> {
-  return normalizeModelCompat(normalizeOpenAICodexTransport(params));
+  return normalizeResolvedProviderModel(params);
 }
 
 export { buildModelAliasLines };
